requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests==2.27.1
python-dotenv==0.19.2
google-cloud-core
customerio >= 1.0